+++ webserver/src/main/java/com/jameswu/demo/repository/OrderDetailRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderDetailRepository extends CrudRepository<OrderDetail, Long> {}
+public interface OrderDetailRepository extends CrudRepository<OrderDetail, Integer> {}
