+++ webserver/src/main/java/com/jameswu/demo/repository/UserProfileRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserProfileRepository extends CrudRepository<UserProfile, Long> {
+public interface UserProfileRepository extends CrudRepository<UserProfile, Integer> {
     Page<UserProfile> findAll(Pageable pageable);
 }
